requirements.txt
pandas
numpy
matplotlib
seaborn
tabula-py
python-pptx
flask
streamlit
pytest